Refactor new user to use authentication tokens.
