+++ app/new-user/new-user.service.ts
@@ -2,41 +2,48 @@ import { Injectable } from "@angular/core";
 import { Angular2Apollo } from "angular2-apollo";
 import gql from "graphql-tag";
 
-import { ConnectivityService }from "../../nativescript-services";
-import { User } from "./";
+import { ConnectivityService } from "../nativescript-services";
+import { LoginService, User } from "../shared";
 
 @Injectable()
-export class UserService {
-  private createUserQuery = gql`
-    mutation newUser($user: AutoUser!) {
-      createAutoUser(user: $user) {
-        _id
-      }
-    }
-  `;
-
+export class NewUserService {
   constructor(
     private apollo: Angular2Apollo,
-    private connectivityService: ConnectivityService
+    private connectivityService: ConnectivityService,
+    private login: LoginService
   ) { }
 
   /**
    * Create new user with type player.
    * @param user The user.
    */
-  public newUser(user: User): Promise<User> {
+  public newUser(user: User): Promise<User | void> {
+    const token = this.login.getToken();
+    if (!token) {
+      return Promise.reject("No authentication token");
+    }
+
     if (!this.connectivityService.isOnline()) {
       this.storeForLaterPush(user);
       return Promise.resolve(user);
     }
 
-    return this.apollo
+    return (this.apollo
       .mutate({
-        mutation: this.createUserQuery,
+        mutation: gql`
+          mutation newUser($token: String!, $user: AutoUser!) {
+            viewer(token: $token) {
+              createAutoUser(user: $user) {
+                _id
+              }
+            }
+          }
+        `,
         variables: {
+          token,
           user,
         },
-      })
+      }) as Promise<void>)
       .catch((error) => {
         console.error("UserService.newUser(): Error:", error);
         return Promise.reject(error);
