+++ app/shared/index.ts
@@ -4,5 +4,5 @@ export * from "./session";
 export * from "./persistence";
 export * from "./util";
 export * from "./login";
-export * from "./user";
 export * from "./measurement.model";
+export * from "./user.model";
