--- app/shared/user/index.ts
@@ -1,2 +1,0 @@
-export * from "./user.model";
-export * from "./user.service";
