+++ package.json
@@ -76,7 +76,7 @@
     "rimraf": "2.5.4",
     "tslint": "3.15.1",
     "typed-graphql": "1.0.2",
-    "typescript": "2.0.7",
+    "typescript": "2.0.8",
     "webpack": "1.13.3"
   }
 }
